\name{GWASTools-defunct}
\alias{GWASTools-defunct}
\alias{ncdfCreate}
\alias{ncdfAddData}
\alias{ncdfAddIntensity}
\alias{ncdfCheckGenotype}
\alias{ncdfCheckIntensity}
\alias{ncdfSetMissingGenotypes}
\alias{gdsSetMissingGenotypes}
\alias{ncdfImputedDosage}
\alias{gdsImputedDosage}
\alias{gdsCheckImputedDosage}
\title{Defunct Functions in Package GWASTools}

\description{
These functions are no longer part of GWASTools as they are no longer needed.
}

\details{
  \code{pedigreeClean} has been replaced by \code{\link{pedigreeCheck}}.
  
  \code{pedigreeFindDuplicates} has been replaced by \code{\link{pedigreeCheck}}.

    \code{ncdfCreate} has been replaced by  \code{\link{createDataFile}}.

    \code{ncdfAddData} has been replaced by  \code{\link{createDataFile}}.

    \code{ncdfAddIntensity} has been replaced by  \code{\link{createAffyIntensityFile}}.

    \code{ncdfCheckGenotype} has been replaced by  \code{\link{checkGenotypeFile}}.

    \code{ncdfCheckIntensity} has been replaced by  \code{\link{checkIntensityFile}}.

    \code{ncdfSetMissingGenotypes} has been replaced by  \code{\link{setMissingGenotypes}}.

    \code{gdsSetMissingGenotypes} has been replaced by  \code{\link{setMissingGenotypes}}.

    \code{ncdfImputedDosage} has been replaced by  \code{\link{imputedDosageFile}}.

    \code{gdsImputedDosage} has been replaced by  \code{\link{imputedDosageFile}}.

    \code{gdsCheckImputedDosage} has been replaced by  \code{\link{checkImputedDosageFile}}.
}
