\name{GWASTools-deprecated}
\alias{GWASTools-deprecated}
\title{Deprecated functions in package \sQuote{GWASTools}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{GWASTools} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{ncdfCreate: \code{\link{createDataFile}}}
    \item{ncdfAddData: \code{\link{createDataFile}}}
    \item{ncdfAddIntensityData: \code{\link{createAffyIntensityFile}}}
    \item{ncdfCheckGenotype: \code{\link{checkGenotypeFile}}}
    \item{ncdfCheckIntensity: \code{\link{checkIntensityFile}}}
    \item{ncdfSetMissingGenotypes: \code{\link{setMissingGenotypes}}}
    \item{gdsSetMissingGenotypes: \code{\link{setMissingGenotypes}}}
    \item{ncdfImputedDosage: \code{\link{imputedDosageFile}}}
    \item{gdsImputedDosage: \code{\link{imputedDosageFile}}}
    \item{gdsCheckImputedDosage: \code{\link{checkImputedDosageFile}}}

  }
}
