% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expected_p.R
\name{get_expected_p}
\alias{get_expected_p}
\title{Get expected P-value based on uniform quantiles}
\usage{
get_expected_p(observed_p, na.rm = FALSE)
}
\arguments{
\item{observed_p}{float for current p-values}

\item{na.rm}{Should NAs be discarded?}
}
\description{
Get expected P-value based on uniform quantiles
}
