% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_int.R
\name{get_int}
\alias{get_int}
\title{inverse normalization transformation (INT)}
\usage{
get_int(x)
}
\arguments{
\item{x}{numerics, typically phenotype values.}
}
\value{
a vector of inverse rank noramlized phenotype values.
}
\description{
inverse normalization transformation (INT)
}
