% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rint.R
\name{get_rint}
\alias{get_rint}
\title{inverse normalization transformation (RINT)}
\usage{
get_rint(x, k = 3/8)
}
\arguments{
\item{x}{numerics, typically phenotype values.}

\item{k}{adjustable offset" (default: Blom offset = 3/8).}
}
\value{
a vector of inverse rank noramlized phenotype values.
}
\description{
inverse normalization transformation (RINT)
}
\references{
\url{https://cran.r-project.org/web/packages/RNOmni/vignettes/RNOmni.html#inverse-normal-transformation}
}
