% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_markers.R
\name{order_markers}
\alias{order_markers}
\title{Order markers by position}
\usage{
order_markers(x, rm.dup = TRUE)
}
\arguments{
\item{x}{vector of markers to be ordered}
}
\value{
a vector of oredered markers.
}
\description{
Order markers by position
}
