% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualize_trait.R
\name{residualize_trait}
\alias{residualize_trait}
\title{residualize trait}
\usage{
residualize_trait(trait, data, covars = c("age", "age2", "ukbb.centre"))
}
\arguments{
\item{trait}{string, trait of interest.}

\item{data}{phenotype data}

\item{covars}{vector of fixed effects to be included}
}
\value{
a vector of residuals
}
\description{
fit fixed effects and residualize the trait
}
