% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualize_trait_by_sex.R
\name{residualize_trait_by_sex}
\alias{residualize_trait_by_sex}
\title{residualize trait seperately for sexes}
\usage{
residualize_trait_by_sex(
  trait,
  data,
  sex = "sex",
  covars = c("age", "age2", "sex", "sex_age")
)
}
\arguments{
\item{trait}{string, trait of interest.}

\item{data}{phenotype data}

\item{sex}{string, column in data that contains sex information (expect 1 or 0)}

\item{covars}{vector of fixed effects to be included}
}
\value{
a vector of residuals
}
\description{
fit fixed effects and residualize the trait
}
\references{
From workflow in \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6488973/}
}
