\name{GWASTools-defunct}
\alias{GWASTools-defunct}
\title{Defunct Functions in Package GWASTools}

\description{
These functions are no longer part of GWASTools as they are no longer needed.
}

\details{
  \code{pedigreeClean} has been replaced by \code{\link{pedigreeCheck}}.
  
  \code{pedigreeFindDuplicates} has been replaced by \code{\link{pedigreeCheck}}.
}
