\name{GWASTools-defunct}
\alias{GWASTools-defunct}
\alias{ncdfCreate}
\alias{ncdfAddData}
\alias{ncdfAddIntensity}
\alias{ncdfCheckGenotype}
\alias{ncdfCheckIntensity}
\alias{ncdfSetMissingGenotypes}
\alias{gdsSetMissingGenotypes}
\alias{ncdfImputedDosage}
\alias{gdsImputedDosage}
\alias{gdsCheckImputedDosage}
\alias{assocTestRegression}
\alias{assocTestCPH}
\alias{assocTestFisherExact}
\alias{gwasExactHW}

\title{Defunct Functions in Package \sQuote{GWASTools}}

\description{
  These functions are defunct and no longer available.
}

\details{
  The following functions are defunct; use the replacement indicated below:
  \itemize{
    \item{\code{pedigreeClean}: \code{\link{pedigreeCheck}}}
    \item{\code{pedigreeFindDuplicates}: \code{\link{pedigreeCheck}}}
    \item{\code{ncdfCreate}: \code{\link{createDataFile}}}
    \item{\code{ncdfAddData}: \code{\link{createDataFile}}}
    \item{\code{ncdfAddIntensity}: \code{\link{createAffyIntensityFile}}}
    \item{\code{ncdfCheckGenotype}: \code{\link{checkGenotypeFile}}}
    \item{\code{ncdfCheckIntensity}: \code{\link{checkIntensityFile}}}
    \item{\code{ncdfSetMissingGenotypes}: \code{\link{setMissingGenotypes}}}
    \item{\code{gdsSetMissingGenotypes}: \code{\link{setMissingGenotypes}}}
    \item{\code{ncdfImputedDosage}: \code{\link{imputedDosageFile}}}
    \item{\code{gdsImputedDosage}: \code{\link{imputedDosageFile}}}
    \item{\code{gdsCheckImputedDosage}: \code{\link{checkImputedDosageFile}}}
    \item{\code{assocTestRegression}: \code{\link{assocRegression}}}
    \item{\code{assocTestCPH}: \code{\link{assocCoxPH}}}
    \item{\code{assocTestFisherExact}: \code{\link{batchFisherTest}}}
    \item{\code{gwasExactHW}: \code{\link{exactHWE}}}
  }   
}
