\name{GWASTools-deprecated}
\alias{GWASTools-deprecated}
\title{Deprecated Functions in Package GWASTools}

\description{
These functions are provided for compatibility with older versions of R only, and may be defunct as soon as the next release.
}

\details{
  \code{\link{pedigreeClean-deprecated}}
  
  \code{\link{pedigreeFindDuplicates-deprecated}}
}
