\name{GWASTools-deprecated}
\alias{GWASTools-deprecated}
\title{Deprecated functions in package \sQuote{GWASTools}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{GWASTools} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{plinkToNcdf: \code{\link[SNPRelate]{snpgdsBED2GDS}}}
    \item{convertVcfGds: \code{\link[SNPRelate]{snpgdsVCF2GDS}}}

  }
}
