\name{GWASTools-deprecated}
\alias{GWASTools-deprecated}
\title{Deprecated functions in package \sQuote{GWASTools}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{GWASTools} only, and will be defunct at the next release.
}

\details{
  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{
    \item{\code{assocTestRegression}: \code{\link{assocRegression}}}
    \item{\code{assocTestCPH}: \code{\link{assocCoxPH}}}
    \item{\code{assocTestRegression}: \code{\link{batchFisherTest}}}
    \item{\code{gwasExactHW}: \code{\link{exactHWE}}}
  }
}
